fix loading product by remarks

+++ lib/data/model/popular_product_item.dart
@@ -10,7 +10,7 @@ class ProductItem {
   String? discountPrice;
   String? image;
   int? stock;
-  double? star;
+  int? star;
   String? remark;
   int? categoryId;
   int? brandId;
